+++ src/BabylonCpp/include/babylon/tools/optimization/hardware_scaling_optimization.h
@@ -19,7 +19,7 @@ public:
   int maximumScale;
 
 private:
-  float _currentScale;
+  int _currentScale;
 
 }; // end of class SceneOptimization
